Move AppLocktType to separate file
