+++ Source/Model/FeatureConfig/AppLock/AppLockController.swift
@@ -21,84 +21,6 @@ import LocalAuthentication
 
 private let zmLog = ZMSLog(tag: "AppLockController")
 
-public protocol AppLockType {
-
-    var delegate: AppLockDelegate? { get set }
-
-    /// Whether the app lock feature is availble to the user.
-
-    var isAvailable: Bool { get }
-
-    /// Whether the app lock on.
-
-    var isActive: Bool { get set }
-
-    /// Whether the app lock is mandatorily active.
-
-    var isForced: Bool { get }
-
-    /// The maximum number of seconds allowed in the background before the
-    /// authentication is required.
-
-    var timeout: UInt { get }
-
-    /// Whether the app lock is currently locked.
-
-    var isLocked: Bool { get }
-
-
-    // TODO: revisit this. It's purpose is to communicate that
-    // the preferred passcode is custom not device.
-
-    var requiresBiometrics: Bool { get }
-
-    // TODO: Rename to "isCustomPasscodeSet"
-    /// Whether a custom passcode has been set.
-
-    var isCustomPasscodeNotSet: Bool { get }
-
-    /// Whether the user needs to be informed about configuration changes.
-
-    var needsToNotifyUser: Bool { get set }
-
-    /// Open the app lock.
-
-    func open() throws
-
-    // TODO: document
-
-    func evaluateAuthentication(passcodePreference: AppLockPasscodePreference,
-                                description: String,
-                                context: LAContextProtocol,
-                                callback: @escaping (AppLockController.AuthenticationResult, LAContextProtocol) -> Void)
-
-    /// Delete the stored passcode.
-
-    func deletePasscode() throws
-
-    /// Update the stored passcode.
-
-    func updatePasscode(_ passcode: String) throws
-
-    // TODO: Document
-
-    func evaluateAuthentication(customPasscode: String) -> AppLockController.AuthenticationResult
-
-}
-
-public extension AppLockType {
-
-    func evaluateAuthentication(passcodePreference: AppLockPasscodePreference,
-                                description: String,
-                                callback: @escaping (AppLockController.AuthenticationResult, LAContextProtocol) -> Void) {
-
-        evaluateAuthentication(passcodePreference: passcodePreference,
-                               description: description,
-                               context: LAContext(),
-                               callback: callback)
-    }
-}
-
 public final class AppLockController: AppLockType {
 
     // MARK: - Properties
